// src/API.ts
export type Platform = {
  __typename: "Platform",
  id: string,
  name?: string | null,
  value?: string | null,
  Products?: ModelProductConnection | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelProductConnection = {
  __typename: "ModelProductConnection",
  items:  Array<Product | null >,
  nextToken?: string | null,
};

export type Product = {
  __typename: "Product",
  id: string,
  name?: string | null,
  isSold?: boolean | null,
  price?: number | null,
  image?: string | null,
  platformID: string,
  genreID: string,
  Platform?: Platform | null,
  Genre?: Genre | null,
  createdAt: string,
  updatedAt: string,
};

export type Genre = {
  __typename: "Genre",
  id: string,
  name?: string | null,
  value?: string | null,
  Products?: ModelProductConnection | null,
  createdAt: string,
  updatedAt: string,
};